Crate near_vm_logic

Re-exports

pub use types::ReturnData;

pub use gas_counter::with_ext_cost_counter;

Modules

gas_counter

mocks

profile

serde_with

types

Structs

ExtCostsConfig

VMConfig

VMContext
Context for the contract execution.

VMLimitConfig
Describes limits for VM and Runtime.

VMLogic

VMOutcome

Enums

ActionCosts

ExtCosts
Strongly-typed representation of the fees for counting.

HostError

VMKind

VMLogicError

Traits

External
An external blockchain interface for the Runtime logic

MemoryLike
An abstraction over the memory of the smart contract.

ValuePtr
Logical pointer to a value in storage. Allows getting value length before getting the value itself. This is needed so that runtime can charge gas before accessing a potentially large value.

Type Definitions

ProtocolVersion